move ulogger API to engine, cover with tests
