+++ .rspec
@@ -1,2 +1,2 @@
---require spec_helper
 --format documentation
+--pattern "spec/**/*_spec.rb,engines/**/*_spec.rb"
