+++ config/application.rb
@@ -18,6 +18,13 @@ require "action_cable/engine"
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 
+# Load available engines
+Dir.glob("#{__dir__}/../engines/*").select { |i| File.directory?(i) }.each do |dir|
+  engine_name = File.basename(dir)
+  filename = File.expand_path(File.join(dir, "lib", "#{engine_name}.rb"))
+  require_relative(filename) if File.exist?(filename)
+end
+
 module Mapforge
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
