+++ config/routes.rb
@@ -28,10 +28,5 @@ Rails.application.routes.draw do
   # Defines the root path route ("/")
   root "frontpage#index"
 
-  post "/ulogger/client/index.php", to: "api/ulogger#auth", constraints: lambda {
- |request| request.request_parameters["action"] == "auth" }
-  post "/ulogger/client/index.php", to: "api/ulogger#addtrack", constraints: lambda {
- |request| request.request_parameters["action"] == "addtrack" }
-  post "/ulogger/client/index.php", to: "api/ulogger#addpos", constraints: lambda {
- |request| request.request_parameters["action"] == "addpos" }
+  mount Ulogger::Engine, at: "/ulogger"
 end
